:sparkles: Legger på feature toggling på sletting av søknad og tilhørende vedlegg

+++ context/soknadcontext/actions.tsx
@@ -90,10 +90,13 @@ export function setSoknadStateFraProps(
 export async function slettLagretSoknadState(
   dispatch: Dispatch<SoknadAction>,
   state: SoknadContextState,
+  brukerMellomLagretSøknadFraAApInnsending: boolean,
 ) {
-  const deleteResponse = await fetch(`/aap/soknad/api/buckets/slett/?type=${state.type}`, {
-    method: 'DELETE',
-  });
+  const deleteResponse = brukerMellomLagretSøknadFraAApInnsending
+    ? await fetch('/aap/soknad/api/mellomlagring/slett')
+    : await fetch(`/aap/soknad/api/buckets/slett/?type=${state.type}`, {
+        method: 'DELETE',
+      });
   return !!deleteResponse?.ok;
 }
 
@@ -148,12 +151,21 @@ export const getVedleggUuidsFromSoknad = (søknad?: Soknad) => {
     .map((vedlegg) => vedlegg?.vedleggId);
 };
 
-export const deleteOpplastedeVedlegg = async (søknad?: Soknad) => {
+export const deleteOpplastedeVedlegg = async (
+  brukMellomlagretSøknadFraInnsending: boolean,
+  søknad?: Soknad,
+) => {
   const vedleggUuids = getVedleggUuidsFromSoknad(søknad);
   if (vedleggUuids.length > 0) {
-    const commaSeparatedUuids = vedleggUuids.join(',');
-    await fetch(`/aap/soknad/api/vedlegg/slett/?uuids=${commaSeparatedUuids}`, {
-      method: 'DELETE',
-    });
+    if (brukMellomlagretSøknadFraInnsending) {
+      for (const vedlegg of vedleggUuids) {
+        await fetch(`/aap/soknad/api/vedlegginnsending/slett/?uuid=${vedlegg}`);
+      }
+    } else {
+      const commaSeparatedUuids = vedleggUuids.join(',');
+      await fetch(`/aap/soknad/api/vedlegg/slett/?uuids=${commaSeparatedUuids}`, {
+        method: 'DELETE',
+      });
+    }
   }
 };
